- arreglos en los permisos

+++ US/views.py
@@ -1,4 +1,5 @@
 from django.contrib.auth.decorators import login_required, permission_required
+from django.core.exceptions import PermissionDenied
 from django.db.models import Q
 from django.shortcuts import render
 from django.http import HttpResponseRedirect
@@ -57,6 +58,58 @@ def create_tipo(request):
         form = TipoUSForm()
 
     return render(request, 'tipo_us_create.html', {'form': form}, )
+
+
+@login_required(None, 'login', '/login/')
+@permission_required('tipous.delete_tipous', raise_exception=True)
+def delete_tipo_us(request, pk):
+    """
+    Busca el tipo de us con pk igual al que es parametro y elimina si no contiene ningun US.
+    Parametros: recibe el request y el pk del tipo a eliminar.
+    Retorna: Redireccion a lista de tipos.
+    """
+    try:
+        tipo_us = TipoUS.objects.get(pk=pk)
+    except:
+        return HttpResponseRedirect('/us/tipo/')
+    if not tipo_us.uss.all():
+        tipo_us.delete()
+
+    return HttpResponseRedirect('/us/tipo/')
+
+
+@login_required(None, 'login', '/login/')
+@permission_required('tipous.change_tipous', raise_exception=True)
+def update_tipo_us(request, pk):
+    """
+        Funcion para actualizar tipo de US utilizando el form TipoUSForm.
+        Recibe en el request el form completado, o displaya uno con los datos previos del tipo en
+        caso de que no se llame a post. Controla la validez del form antes de guardarlo como un tipo
+         nuevo en la base de datos.
+        Parametros: Recibe el request y el pk del tipo a editar.
+        Retorna:
+        -El render del template tipo_us_create.html en caso de form vacio o invalido.
+        -Redireccion a lista de tipos si el form es valido
+
+    """
+    try:
+        tipo_us = TipoUS.objects.get(pk=pk)
+    except:
+        return HttpResponseRedirect('/us/tipo/')
+
+    if request.method == 'POST':
+        form = TipoUSForm(request.POST)
+        if form.is_valid():
+            cd = form.cleaned_data
+            tipo_us.nombre = cd['nombre']
+            tipo_us.save()
+            return HttpResponseRedirect('/us/tipo/' + str(tipo_us.id))
+    else:
+        tipo_us = TipoUS.objects.get(pk=pk)
+        form = TipoUSForm(initial={'nombre': tipo_us.nombre})
+
+    return render(request, 'tipo_us_create.html', {'form': form, 'tipo_us': tipo_us})
+
 
 
 @login_required(None, 'login', '/login/')
@@ -128,6 +181,9 @@ def create_us(request, pk):
     except:
         print 'Error'
 
+    permisos = proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename',
+                                                                                        flat=True)
+
     form = USForm(request.POST or None)
     equipos = proyecto.equipos.all()
     usuarios = Usuario.objects.filter(Q(equipos__in=equipos) | Q(lider=proyecto)).distinct()
@@ -171,55 +227,6 @@ def delete_actividad(request, pk):
     return HttpResponseRedirect('/us/tipo/' + str(tipo_us.id))
 
 
-@login_required(None, 'login', '/login/')
-@permission_required('tipous.delete_tipous', raise_exception=True)
-def delete_tipo_us(request, pk):
-    """
-    Busca el tipo de us con pk igual al que es parametro y elimina si no contiene ningun US.
-    Parametros: recibe el request y el pk del tipo a eliminar.
-    Retorna: Redireccion a lista de tipos.
-    """
-    try:
-        tipo_us = TipoUS.objects.get(pk=pk)
-    except:
-        return HttpResponseRedirect('/us/tipo/')
-    if not tipo_us.uss.all():
-        tipo_us.delete()
-
-    return HttpResponseRedirect('/us/tipo/')
-
-
-@login_required(None, 'login', '/login/')
-def update_tipo_us(request, pk):
-    """
-        Funcion para actualizar tipo de US utilizando el form TipoUSForm.
-        Recibe en el request el form completado, o displaya uno con los datos previos del tipo en
-        caso de que no se llame a post. Controla la validez del form antes de guardarlo como un tipo
-         nuevo en la base de datos.
-        Parametros: Recibe el request y el pk del tipo a editar.
-        Retorna:
-        -El render del template tipo_us_create.html en caso de form vacio o invalido.
-        -Redireccion a lista de tipos si el form es valido
-
-    """
-    try:
-        tipo_us = TipoUS.objects.get(pk=pk)
-    except:
-        return HttpResponseRedirect('/us/tipo/')
-
-    if request.method == 'POST':
-        form = TipoUSForm(request.POST)
-        if form.is_valid():
-            cd = form.cleaned_data
-            tipo_us.nombre = cd['nombre']
-            tipo_us.save()
-            return HttpResponseRedirect('/us/tipo/' + str(tipo_us.id))
-    else:
-        tipo_us = TipoUS.objects.get(pk=pk)
-        form = TipoUSForm(initial={'nombre': tipo_us.nombre})
-
-    return render(request, 'tipo_us_create.html', {'form': form, 'tipo_us': tipo_us})
-
 
 @login_required(None, 'login', '/login/')
 def detail_us(request, pk):
@@ -233,8 +240,12 @@ def detail_us(request, pk):
 
     permisos = us.proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename',
                                                                                            flat=True)
+    if ('view_us' in permisos or us.usuario_asignado == request.user or request.user.is_staff):
+        return render(request, 'us_detail.html', {'object': us, 'permisos': permisos})
+    else:
+        raise PermissionDenied
 
-    return render(request, 'us_detail.html', {'object': us, 'permisos': permisos})
+
 
 
 @login_required(None, 'login', '/login/')
@@ -244,12 +255,19 @@ def delete_us(request, pk):
     Parametros: recibe el request y el pk del us a eliminar.
     Retorna: Redireccion a ldetalle del proyecto.
     """
+
     try:
         us = US.objects.get(pk=pk)
     except:
         return HttpResponseRedirect('/proyecto/')
-    proyecto = us.proyecto
-    us.delete()
+
+    permisos = us.proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename',flat=True)
+    if ('delete_us' in permisos or request.user.is_staff):
+        proyecto = us.proyecto
+        us.delete()
+    else:
+        raise PermissionDenied
+
 
     return HttpResponseRedirect('/proyecto/' + str(proyecto.id))
 
@@ -271,34 +289,41 @@ def update_us(request, pk):
         us = US.objects.get(pk=pk)
     except:
         return HttpResponseRedirect('/proyecto/')
+    permisos = us.proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename',
+                                                                                           flat=True)
+    if ('update_us' in permisos or request.user.is_staff):
+        equipos = us.proyecto.equipos.all()
+        usuarios = Usuario.objects.filter(Q(equipos__in=equipos) | Q(lider=us.proyecto)).distinct()
+        if request.method == 'POST':
+            form = USForm(request.POST)
+            form.fields["usuario_asignado"].queryset = usuarios
+            if form.is_valid():
+                us.descripcion_corta = form.cleaned_data['descripcion_corta']
+                us.descripcion_larga = form.cleaned_data['descripcion_larga']
+                us.tiempo_planificado = form.cleaned_data['tiempo_planificado']
+                us.valor_negocio = form.cleaned_data['valor_negocio']
+                us.urgencia = form.cleaned_data['urgencia']
+                us.usuario_asignado = form.cleaned_data['usuario_asignado']
+                us.tipoUS = form.cleaned_data['tipoUS']
+                us.save()
+                return HttpResponseRedirect('/us/us/' + str(us.id))
+        else:
+            form = USForm(initial={'descripcion_corta': us.descripcion_corta,
+                                   'descripcion_larga': us.descripcion_larga,
+                                   'tiempo_planificado': us.tiempo_planificado,
+                                   'valor_negocio': us.valor_negocio,
+                                   'urgencia': us.urgencia,
+                                   'usuario_asignado': us.usuario_asignado,
+                                   'tipoUS': us.tipoUS})
 
-    equipos = us.proyecto.equipos.all()
-    usuarios = Usuario.objects.filter(Q(equipos__in=equipos) | Q(lider=us.proyecto)).distinct()
-    if request.method == 'POST':
-        form = USForm(request.POST)
         form.fields["usuario_asignado"].queryset = usuarios
-        if form.is_valid():
-            us.descripcion_corta = form.cleaned_data['descripcion_corta']
-            us.descripcion_larga = form.cleaned_data['descripcion_larga']
-            us.tiempo_planificado = form.cleaned_data['tiempo_planificado']
-            us.valor_negocio = form.cleaned_data['valor_negocio']
-            us.urgencia = form.cleaned_data['urgencia']
-            us.usuario_asignado = form.cleaned_data['usuario_asignado']
-            us.tipoUS = form.cleaned_data['tipoUS']
-            us.save()
-            return HttpResponseRedirect('/us/us/' + str(us.id))
+
+        return render(request, 'us_create.html', {'form': form, 'us': us}, )
     else:
-        form = USForm(initial={'descripcion_corta': us.descripcion_corta,
-                               'descripcion_larga': us.descripcion_larga,
-                               'tiempo_planificado': us.tiempo_planificado,
-                               'valor_negocio': us.valor_negocio,
-                               'urgencia': us.urgencia,
-                               'usuario_asignado': us.usuario_asignado,
-                               'tipoUS': us.tipoUS})
+        raise PermissionDenied
 
-    form.fields["usuario_asignado"].queryset = usuarios
 
-    return render(request, 'us_create.html', {'form': form, 'us': us}, )
+
 
 
 @login_required(None, 'login', '/login/')
@@ -318,30 +343,35 @@ def cambiar_actividad(request, pk):
     except:
         return HttpResponseRedirect('/proyecto/')
 
-    actividades = us.tipoUS.actividades.all()
+    permisos = us.proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename',
+                                                                                           flat=True)
+    if ('update_us' in permisos or request.user.is_staff):
+        actividades = us.tipoUS.actividades.all()
 
-    if request.method == 'POST':
-        form = CambiarActividadForm(request.POST)
-        form.fields['actividad'].queryset = actividades
-        form_estado = CambiarEstadoActividadForm(request.POST)
-        finalizado = request.POST.get('finalizado')
-        if finalizado == 'Si':
-            us.finalizado = True
+        if request.method == 'POST':
+            form = CambiarActividadForm(request.POST)
+            form.fields['actividad'].queryset = actividades
+            form_estado = CambiarEstadoActividadForm(request.POST)
+            finalizado = request.POST.get('finalizado')
+            if finalizado == 'Si':
+                us.finalizado = True
+            else:
+                us.finalizado = False
+            if form.is_valid() and form_estado.is_valid():
+                us.estado_actividad = form_estado.cleaned_data['estado_actividad']
+                us.actividad = form.cleaned_data['actividad']
+
+                us.save()
+                return HttpResponseRedirect('/us/us/' + str(us.id))
         else:
-            us.finalizado = False
-        if form.is_valid() and form_estado.is_valid():
-            us.estado_actividad = form_estado.cleaned_data['estado_actividad']
-            us.actividad = form.cleaned_data['actividad']
+            form = CambiarActividadForm(initial={'actividad': us.actividad})
+            form_estado = CambiarEstadoActividadForm(initial={'estado_actividad': us.estado_actividad})
+            fin = us.finalizado
+        form.fields['actividad'].queryset = actividades
 
-            us.save()
-            return HttpResponseRedirect('/us/us/' + str(us.id))
+        return render(request, 'cambiar_actividad.html', {'form': form, 'us': us, 'form_estado': form_estado, 'fin': fin})
     else:
-        form = CambiarActividadForm(initial={'actividad': us.actividad})
-        form_estado = CambiarEstadoActividadForm(initial={'estado_actividad': us.estado_actividad})
-        fin = us.finalizado
-    form.fields['actividad'].queryset = actividades
-
-    return render(request, 'cambiar_actividad.html', {'form': form, 'us': us, 'form_estado': form_estado, 'fin': fin})
+        raise PermissionDenied
 
 
 @login_required(None, 'login', '/login/')
@@ -360,50 +390,39 @@ def cambiar_estado_actividad(request, pk):
         us = US.objects.get(pk=pk)
     except:
         return HttpResponseRedirect('/proyecto/')
-    if (us.estado_actividad == 'TOD'):
-        us.estado_actividad = 'DOI'
-        us.save()
-        return HttpResponseRedirect('/us/us/' + str(us.id))
-    elif (us.estado_actividad == 'DOI'):
-        us.estado_actividad = 'DON'
-        us.save()
-        return HttpResponseRedirect('/us/us/' + str(us.id))
-    else:
-        """
-        actividad_nueva=Actividades.objects.filter(tipoUS__pk=pk).filter(numero__gt=us.actividad.numero).order_by('numero')[0:1]
-        if actividad_nueva:
-            us.actividad=actividad_nueva[0]
-            us.estado_actividad = 'TOD'
+
+    permisos = us.proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename',
+                                                                                               flat=True)
+    if ('update_us' in permisos or request.user.is_staff):
+        if (us.estado_actividad == 'TOD'):
+            us.estado_actividad = 'DOI'
             us.save()
             return HttpResponseRedirect('/us/us/' + str(us.id))
-        else:"""
-
-        return render(request, 'ultimo_estado.html', {'pk': pk})
-
-    """
-    if request.method == 'POST':
-        form = CambiarEstadoActividadForm(request.POST)
-        if form.is_valid():
-            us.estado_actividad = form.cleaned_data['estado_actividad']
+        elif (us.estado_actividad == 'DOI'):
+            us.estado_actividad = 'DON'
             us.save()
             return HttpResponseRedirect('/us/us/' + str(us.id))
-
+        else:
+            return render(request, 'ultimo_estado.html', {'pk': pk})
     else:
-        form = CambiarEstadoActividadForm(initial={'estado_actividad': us.estado_actividad})
-
-    return render(request, 'cambiar_estado_actividad.html', {'form': form, 'us': us})
-    """
+        raise PermissionDenied
 
 
 def repriorizar(request, pk):
+
     try:
         us = US.objects.get(pk=pk)
+    except:
+        return HttpResponseRedirect('/proyecto/')
+
+    permisos = us.proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename',flat=True)
+    if ('update_us' in permisos or request.user.is_staff):
         us.estado_actividad = 'TOD'
         us.urgencia = 5
         us.valor_negocio = 5
         us.sprint = None
         us.save()
-    except:
-        return HttpResponseRedirect('/proyecto/')
+    else:
+        raise PermissionDenied
 
-    return HttpResponseRedirect('/proyecto/')
+    return HttpResponseRedirect('/proyecto/'+us.proyecto.id+'/')
